api/requirements.txt
pytest
uvicorn
fastapi
sqlalchemy
python-gitlab